Rename `soft_contacts` module to `soft`
